Remove layouts from every subfolder and setup one root layout again.
- Need to draw top bars in each page rather than in the layout, since the root layout is meant to stay the same for each view and the top bar definitely changes in different views.

- Will just use layout to set the background color and stuff, top bars in each of their own pages.
