+++ package.json
@@ -16,7 +16,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "lucide-react": "^0.378.0",
-    "next": "14.2.3",
+    "next": "^14.2.4",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-dom": "^18",
