--- src/app/home/layout.tsx
@@ -1,19 +1,0 @@
-import "./../styles/globals.css";
-import React from "react";
-
-import TopBar from "@/components/TopBar";
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body className={"bg-background"}>
-        <TopBar />
-      {children}
-      </body>
-    </html>
-  );
-}
