+++ src/app/weld/page.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react';
 
-
 export default async function Home() {
+  
   return (
     <main>
     </main>
